revamped transformer, migrated historical data source class
