+++ src/utils/common-expression-helpers.ts
@@ -94,16 +94,13 @@ export class CommonExpressionHelpers {
     return null;
   };
 
-  doesNotMatchExpression = (
-    regexString: string,
-    val: string | null | undefined
-  ): boolean => {
+  doesNotMatchExpression = (regexString: string, val: string | null | undefined): boolean => {
     if (!val || ['undefined', 'null', ''].includes(val.toString())) {
       return true;
     }
     const pattern = new RegExp(regexString);
-    
-    return !pattern.test(val)
+
+    return !pattern.test(val);
   };
 
   calcBMI = (height: number, weight: number) => {
@@ -455,17 +452,6 @@ export class CommonExpressionHelpers {
   resolve = (lazy: Promise<unknown>) => {
     return Promise.resolve(lazy);
   };
-}
-export class HistoricalDataSourceService {
-  dataSourceMap: Record<string, any> = {};
-
-  putObject(key: string, value: any) {
-    this.dataSourceMap[key] = value;
-  }
-
-  getObject(key: string) {
-    return this.dataSourceMap[key];
-  }
 }
 
 export function registerDependency(node: FormNode, determinant: FormField) {
