+++ src/utils/historical-data-source.ts
@@ -1,0 +1,11 @@
+export class HistoricalDataSourceService {
+  dataSourceMap: Record<string, any> = {};
+
+  putObject(key: string, value: any) {
+    this.dataSourceMap[key] = value;
+  }
+
+  getObject(key: string) {
+    return this.dataSourceMap[key];
+  }
+}
